requirements.txt
PyQt6>=6.4.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
matplotlib>=3.6.0
seaborn>=0.12.0
deep-translator>=1.9.0
joblib>=1.2.0
google-generativeai>=0.8.0
